backend/requirements.txt
# ============================================
# FastAPI Backend Requirements
# Python 3.12 | Railway Production
# ============================================

# ===== Web Framework =====
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1

# ===== Database & ORM =====
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pgvector==0.3.6

# ===== Supabase Client =====
supabase==2.9.0
storage3==0.8.1
postgrest==0.17.2

# ===== AI & Machine Learning =====
openai==1.54.0

# ===== Data Processing =====
numpy==2.1.3

# ===== Image Processing =====
Pillow==11.0.0
PyMuPDF==1.24.14
opencv-python==4.10.0.84
pytesseract==0.3.13

# ===== HTTP Client =====
requests==2.32.3
httpx==0.27.2

# ===== Environment Variables =====
python-dotenv==1.0.1

# ===== Type Checking & Validation =====
typing-extensions==4.12.2

# ===== ASGI Server (Production) =====
gunicorn==23.0.0

# ===== Utilities =====
python-multipart==0.0.20

